refactor: split encoding/encodingread.go — plain reads into separate file
Extract plain-encoding read functions into encodingread_plain.go; split
matching tests into encodingread_plain_test.go.
